requirements.txt
sqlalchemy
python-decouple
psycopg2
pgvector
openai
